requirements.txt
streamlit
python-dotenv
google-generativeai
PyPDF2
Pillow
numpy
pandas
scikit-learn
pydantic
pydantic-settings
motor
bson
certifi
fastapi
uvicorn
gunicorn
python-jose[cryptography]
passlib[bcrypt]
bcrypt
httpx
python-multipart
